chore: create CredentialProvider abstract class
