+++ src/Auth/EnvMomentoTokenProvider.php
@@ -4,19 +4,12 @@ declare(strict_types=1);
 namespace Momento\Auth;
 
 use Momento\Cache\Errors\InvalidArgumentError;
-use function \Momento\Utilities\isNullOrEmpty;
 
 /**
  * Reads and parses a JWT token stored as an environment variable.
  */
-class EnvMomentoTokenProvider implements ICredentialProvider
+class EnvMomentoTokenProvider extends CredentialProvider
 {
-    private string $authToken;
-    private string $controlEndpoint;
-    private string $cacheEndpoint;
-    private ?string $trustedControlEndpointCertificateName = null;
-    private ?string $trustedCacheEndpointCertificateName = null;
-
     public function __construct(
         string  $envVariableName,
         ?string $controlEndpoint = null,
@@ -25,47 +18,7 @@ class EnvMomentoTokenProvider implements ICredentialProvider
         ?string $trustedCacheEndpointCertificateName = null
     )
     {
-        $authToken = getenv($envVariableName);
-        if ($authToken === false || isNullOrEmpty($authToken)) {
-            throw new InvalidArgumentError("Environment variable $envVariableName is empty or null.");
-        }
-        if ($trustedControlEndpointCertificateName xor $trustedCacheEndpointCertificateName) {
-            throw new InvalidArgumentError(
-                "If either of trustedCacheEndpointCertificateName or trustedControlEndpointCertificateName " .
-                "are provided, they must both be."
-            );
-        }
-        $this->authToken = $authToken;
-        $payload = AuthUtils::parseAuthToken($authToken);
-        $this->controlEndpoint = $controlEndpoint ?? $payload->cp;
-        $this->cacheEndpoint = $cacheEndpoint ?? $payload->c;
-        $this->trustedControlEndpointCertificateName = $trustedControlEndpointCertificateName;
-        $this->trustedCacheEndpointCertificateName = $trustedCacheEndpointCertificateName;
-    }
-
-    public function getAuthToken(): string
-    {
-        return $this->authToken;
-    }
-
-    public function getCacheEndpoint(): string
-    {
-        return $this->cacheEndpoint;
-    }
-
-    public function getControlEndpoint(): string
-    {
-        return $this->controlEndpoint;
-    }
-
-    public function getTrustedControlEndpointCertificateName(): string|null
-    {
-        return $this->trustedControlEndpointCertificateName;
-    }
-
-    public function getTrustedCacheEndpointCertificateName(): string|null
-    {
-        return $this->trustedCacheEndpointCertificateName;
+        parent::__construct($envVariableName, $controlEndpoint, $cacheEndpoint, $trustedControlEndpointCertificateName, $trustedCacheEndpointCertificateName);
     }
 
     /**
@@ -74,7 +27,8 @@ class EnvMomentoTokenProvider implements ICredentialProvider
      * @return EnvMomentoTokenProvider
      * @throws InvalidArgumentError
      */
-    public static function fromEnvironmentVariable(string $envVariableName): EnvMomentoTokenProvider {
+    public static function fromEnvironmentVariable(string $envVariableName): EnvMomentoTokenProvider
+    {
         return new EnvMomentoTokenProvider($envVariableName);
     }
 }
