+++ src/Auth/StringMomentoTokenProvider.php
@@ -4,19 +4,12 @@ declare(strict_types=1);
 namespace Momento\Auth;
 
 use Momento\Cache\Errors\InvalidArgumentError;
-use function \Momento\Utilities\isNullOrEmpty;
 
 /**
  * Reads and parses a JWT token stored as a string.
  */
-class StringMomentoTokenProvider implements ICredentialProvider
+class StringMomentoTokenProvider extends CredentialProvider
 {
-    private string $authToken;
-    private string $controlEndpoint;
-    private string $cacheEndpoint;
-    private ?string $trustedControlEndpointCertificateName = null;
-    private ?string $trustedCacheEndpointCertificateName = null;
-
     public function __construct(
         string  $authToken,
         ?string $controlEndpoint = null,
@@ -25,46 +18,7 @@ class StringMomentoTokenProvider implements ICredentialProvider
         ?string $trustedCacheEndpointCertificateName = null
     )
     {
-        if (isNullOrEmpty($authToken)) {
-            throw new InvalidArgumentError("String $authToken is empty or null.");
-        }
-        if ($trustedControlEndpointCertificateName xor $trustedCacheEndpointCertificateName) {
-            throw new InvalidArgumentError(
-                "If either of trustedCacheEndpointCertificateName or trustedControlEndpointCertificateName " .
-                "are provided, they must both be."
-            );
-        }
-        $this->authToken = $authToken;
-        $payload = AuthUtils::parseAuthToken($authToken);
-        $this->controlEndpoint = $controlEndpoint ?? $payload->cp;
-        $this->cacheEndpoint = $cacheEndpoint ?? $payload->c;
-        $this->trustedControlEndpointCertificateName = $trustedControlEndpointCertificateName;
-        $this->trustedCacheEndpointCertificateName = $trustedCacheEndpointCertificateName;
-    }
-
-    public function getAuthToken(): string
-    {
-        return $this->authToken;
-    }
-
-    public function getCacheEndpoint(): string
-    {
-        return $this->cacheEndpoint;
-    }
-
-    public function getControlEndpoint(): string
-    {
-        return $this->controlEndpoint;
-    }
-
-    public function getTrustedControlEndpointCertificateName(): string|null
-    {
-        return $this->trustedControlEndpointCertificateName;
-    }
-
-    public function getTrustedCacheEndpointCertificateName(): string|null
-    {
-        return $this->trustedCacheEndpointCertificateName;
+        parent::__construct($authToken, $controlEndpoint, $cacheEndpoint, $trustedControlEndpointCertificateName, $trustedCacheEndpointCertificateName);
     }
 
     /**
